+++ ts/src/core/loaders/assoc_edge_loader.ts
@@ -55,6 +55,7 @@ function createLoader<T extends AssocEdge>(
   edgeType: string,
   edgeCtr: AssocEdgeConstructor<T>,
   edgeData: AssocEdgeData,
+  context?: Context,
 ) {
   const loaderName = `assocEdgeLoader:${edgeType}`;
   const loaderOptions: DataLoader.Options<ID, T[]> = {
@@ -74,6 +75,7 @@ function createLoader<T extends AssocEdge>(
           id1: keys[0],
           edgeType: edgeType,
           queryOptions: options,
+          context,
           ctr: edgeCtr,
         });
         return [r];
@@ -159,6 +161,7 @@ export class AssocEdgeLoader<T extends AssocEdge> implements Loader<ID, T[]> {
       this.edgeType,
       this.edgeCtr,
       edgeData,
+      this.context,
     );
     return this.loader;
   }