+++ ts/src/core/loaders/query_loader.test.ts
@@ -2,7 +2,8 @@ import { TestContext } from "../../testutils/context/test_context";
 import { setLogLevels } from "../logger";
 import { MockLogs } from "../../testutils/mock_log";
 import { getDefaultLimit } from "../ent";
-import { buildQuery } from "../query_impl";
+import { getContextCacheKey } from "../context";
+import { buildQuery, OrderBy } from "../query_impl";
 import * as clause from "../clause";
 import { Data, EdgeQueryableDataOptions, ID, Loader } from "../base";
 import { setupSqlite, TempDB } from "../../testutils/db/temp_db";
@@ -322,14 +323,16 @@ function commonTests() {
     });
 
     test("multi-ids. with context, offset", async () => {
-      await testMultiQueryDataOffset((options) =>
-        getConfigurableLoader(true, options),
+      await testMultiQueryDataOffset(
+        (options) => getConfigurableLoader(true, options),
+        true,
       );
     });
 
     test("multi-ids. without context, offset", async () => {
-      await testMultiQueryDataOffset((options) =>
-        getConfigurableLoader(false, options),
+      await testMultiQueryDataOffset(
+        (options) => getConfigurableLoader(false, options),
+        false,
       );
     });
 
@@ -717,6 +720,7 @@ function commonTests() {
 
   async function testMultiQueryDataOffset(
     loaderFn: (opts: EdgeQueryableDataOptions) => Loader<ID, Data[]>,
+    expectCacheHit: boolean,
   ) {
     const { m, ids, users } = await createData();
 
@@ -783,10 +787,14 @@ function commonTests() {
     // query again, same data
     // if context, we hit local cache. otherwise, hit db
     expect(edges).toStrictEqual(edges2);
-    verifyMultiCountQueryOffset(ids, m);
+    verifyMultiCountQueryOffset(ids, m, expectCacheHit);
   }
 
-  function verifyMultiCountQueryOffset(ids: ID[], m: Map<ID, FakeEvent[]>) {
+  function verifyMultiCountQueryOffset(
+    ids: ID[],
+    m: Map<ID, FakeEvent[]>,
+    cachehit?: boolean,
+  ) {
     expect(ml.logs.length).toBe(ids.length);
     ml.logs.forEach((log, idx) => {
       let events = m.get(ids[idx]) || [];
@@ -796,6 +804,26 @@ function commonTests() {
         getCompleteClause(ids[idx]),
         clause.Greater("start_time", events[0].startTime.toISOString()),
       );
+      if (cachehit) {
+        const orderby: OrderBy = [
+          {
+            column: "start_time",
+            direction: "ASC",
+          },
+        ];
+        const cacheKey = getContextCacheKey({
+          tableName: "fake_events",
+          fields,
+          clause: cls,
+          orderby,
+          limit: 1,
+        });
+        expect(log).toStrictEqual({
+          "cache-hit": cacheKey,
+          "tableName": "fake_events",
+        });
+        return;
+      }
 
       const expQuery = buildQuery({
         tableName: "fake_events",